requirements.txt
azure-devops>=7.1.0b4
msrest>=0.7.1
platformdirs>=4.2.0
PyYAML>=6.0.1

